💄 Update the dashboard layout
